answers/worker/requirements.txt
Flask==3.0.0
pytest==7.0.1; python_version > "3.0"
# pin pytest to 4.6.11 for Python2.
pytest==7.0.1; python_version < "3.0"
gunicorn==20.1.0
Werkzeug==3.0.1
rasterio
numpy
google-cloud-storage
